actor-model/persistence/dynamodb/persistence/dynamodb_provider.go
package persistence

// ProviderState is an object containing the implementation for the provider
type ProviderState struct {
	// TODO: Add necessary fields
}

// NewProviderState creates a new instance of ProviderState
func NewProviderState() *ProviderState {
	return &ProviderState{
		// TODO: Initialize fields
	}
}

// GetState returns the current state of the provider
func (p *ProviderState) GetState() *ProviderState {
	// TODO: Implement getting the current state
	return p
}

// Restart restarts the provider
func (p *ProviderState) Restart() {
	// TODO: Implement restarting the provider
}

// GetSnapshotInterval returns the snapshot interval
func (p *ProviderState) GetSnapshotInterval() int {
	// TODO: Implement getting the snapshot interval
	return 0
}